+++ src/VaultSharp/V1/AuthMethods/RADIUS/IRADIUSAuthMethod.cs
@@ -1,6 +1,6 @@
 ﻿namespace VaultSharp.V1.AuthMethods.RADIUS
 {
-    public class IRADIUSAuthBackend
+    public class IRADIUSAuthMethod
     {
     }
 }